Fix NullPointer if dir does not exist

+++ src/main/java/eionet/gdem/web/listeners/WebSessionListener.java
@@ -25,6 +25,7 @@ public class WebSessionListener implements HttpSessionListener {
         String id = httpSessionEvent.getSession().getId();
         String sessionDir = Properties.appRootFolder + "/tmpfile/" + id;
         File dir = new File(sessionDir);
+        if ( !dir.exists()) return;
         for(File file: dir.listFiles()) {
             file.delete();
         }
